[generator] implemented new validator fragment
