+++ plugins/org.eclipse.xtend.core/src-gen/org/eclipse/xtend/core/XtendStandaloneSetupGenerated.java
@@ -39,6 +39,5 @@ public class XtendStandaloneSetupGenerated implements ISetup {
 		org.eclipse.xtext.resource.IResourceServiceProvider.Registry.INSTANCE.getExtensionToFactoryMap().put("xtend", serviceProvider);
 		
 		
-		
 	}
 }
