+++ plugins/org.eclipse.xtext.xbase/generator/org/eclipse/xtext/xbase/GenerateXbase.java
@@ -11,7 +11,6 @@ import org.eclipse.xtext.generator.parser.antlr.XtextAntlrGeneratorFragment;
 import org.eclipse.xtext.generator.parser.antlr.XtextAntlrUiGeneratorFragment;
 import org.eclipse.xtext.generator.resourceFactory.ResourceFactoryFragment;
 import org.eclipse.xtext.generator.serializer.SerializerFragment;
-import org.eclipse.xtext.generator.validation.JavaValidatorFragment;
 import org.eclipse.xtext.resource.XtextResourceSet;
 import org.eclipse.xtext.ui.generator.contentAssist.JavaBasedContentAssistFragment;
 import org.eclipse.xtext.ui.generator.labeling.LabelProviderFragment;
@@ -27,6 +26,7 @@ import org.eclipse.xtext.xtext.generator.formatting.Formatter2Fragment2;
 import org.eclipse.xtext.xtext.generator.grammarAccess.GrammarAccessFragment2;
 import org.eclipse.xtext.xtext.generator.scoping.ImportNamespacesScopingFragment2;
 import org.eclipse.xtext.xtext.generator.types.TypesGeneratorFragment2;
+import org.eclipse.xtext.xtext.generator.validation.ValidatorFragment2;
 import org.eclipse.xtext.xtext.generator.xbase.XbaseGeneratorFragment2;
 
 /**
@@ -98,6 +98,7 @@ final class GenerateXbase {
 					setEncoding("ISO-8859-1");
 					setLineDelimiter(lineDelimiter);
 					setFileHeader(fileHeader);
+					setPreferXtendStubs(false);
 				}});
 			}});
 
@@ -134,9 +135,9 @@ final class GenerateXbase {
 				DebugAntlrGeneratorFragment antlrDebug = new DebugAntlrGeneratorFragment();
 				antlrDebug.setOptions(antlrOptions);
 				addFragment(new FragmentAdapter(antlrDebug));
-				JavaValidatorFragment validator = new JavaValidatorFragment();
+				ValidatorFragment2 validator = new ValidatorFragment2();
 				validator.setInheritImplementation(false);
-				addFragment(new FragmentAdapter(validator));
+				addFragment(validator);
 				addFragment(new ImportNamespacesScopingFragment2());
 				addFragment(new TypesGeneratorFragment2());
 				XbaseGeneratorFragment2 xbase = new XbaseGeneratorFragment2();
@@ -177,7 +178,7 @@ final class GenerateXbase {
 				DebugAntlrGeneratorFragment antlrDebug = new DebugAntlrGeneratorFragment();
 				antlrDebug.setOptions(antlrOptions);
 				addFragment(new FragmentAdapter(antlrDebug));
-				addFragment(new FragmentAdapter(new JavaValidatorFragment()));
+				addFragment(new ValidatorFragment2());
 				addFragment(new ImportNamespacesScopingFragment2());
 				addFragment(new TypesGeneratorFragment2());
 				XbaseGeneratorFragment2 xbase = new XbaseGeneratorFragment2();
