+++ plugins/org.eclipse.xtext.xbase/src-gen/org/eclipse/xtext/xbase/annotations/validation/AbstractXbaseWithAnnotationsValidator.java
@@ -5,22 +5,23 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.eclipse.xtext.xbase.validation;
- 
+package org.eclipse.xtext.xbase.annotations.validation;
+
 import java.util.ArrayList;
 import java.util.List;
-
 import org.eclipse.emf.ecore.EPackage;
+import org.eclipse.xtext.xbase.validation.XbaseValidator;
 
-public class AbstractXbaseJavaValidator extends org.eclipse.xtext.validation.AbstractDeclarativeValidator {
-
+public abstract class AbstractXbaseWithAnnotationsValidator extends XbaseValidator {
+	
 	@Override
 	protected List<EPackage> getEPackages() {
-	    List<EPackage> result = new ArrayList<EPackage>();
+	    List<EPackage> result = new ArrayList<EPackage>(super.getEPackages());
+	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/Xtext/Xbase/XAnnotations"));
 	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/xtext/xbase/Xbase"));
 	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/xtext/common/JavaVMTypes"));
 	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/xtext/xbase/Xtype"));
 		return result;
 	}
-
+	
 }
