+++ plugins/org.eclipse.xtext.xbase/src-gen/org/eclipse/xtext/xbase/validation/AbstractXbaseValidator.java
@@ -5,23 +5,22 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.eclipse.xtext.xbase.annotations.validation;
- 
+package org.eclipse.xtext.xbase.validation;
+
 import java.util.ArrayList;
 import java.util.List;
-
 import org.eclipse.emf.ecore.EPackage;
+import org.eclipse.xtext.validation.AbstractDeclarativeValidator;
 
-public class AbstractXbaseWithAnnotationsJavaValidator extends org.eclipse.xtext.xbase.validation.XbaseJavaValidator {
-
+public abstract class AbstractXbaseValidator extends AbstractDeclarativeValidator {
+	
 	@Override
 	protected List<EPackage> getEPackages() {
 	    List<EPackage> result = new ArrayList<EPackage>();
-	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/Xtext/Xbase/XAnnotations"));
 	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/xtext/xbase/Xbase"));
 	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/xtext/common/JavaVMTypes"));
 	    result.add(EPackage.Registry.INSTANCE.getEPackage("http://www.eclipse.org/xtext/xbase/Xtype"));
 		return result;
 	}
-
+	
 }
