+++ plugins/org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/annotations/validation/AbstractXbaseWithAnnotationsJavaValidator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,8 +8,8 @@
 package org.eclipse.xtext.xbase.annotations.validation;
 
 /**
- * @deprecated use {@link XbaseWithAnnotationsValidator} instead
+ * @deprecated use {@link AbstractXbaseWithAnnotationsValidator} instead
  */
 @Deprecated
-public class XbaseWithAnnotationsJavaValidator extends XbaseWithAnnotationsValidator {
+public class AbstractXbaseWithAnnotationsJavaValidator extends AbstractXbaseWithAnnotationsValidator {
 }
