+++ plugins/org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/validation/AbstractXbaseJavaValidator.java
@@ -1,15 +1,15 @@
 /*******************************************************************************
- * Copyright (c) 2011 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.eclipse.xtext.xbase.annotations.validation;
+package org.eclipse.xtext.xbase.validation;
 
 /**
- * @deprecated use {@link XbaseWithAnnotationsValidator} instead
+ * @deprecated use {@link AbstractXbaseValidator} instead
  */
 @Deprecated
-public class XbaseWithAnnotationsJavaValidator extends XbaseWithAnnotationsValidator {
+public class AbstractXbaseJavaValidator extends AbstractXbaseValidator {
 }
